refactor: 상품 상세 내용 컴포넌트 분리 후 페이지 서버 컴포넌트로 변경
