+++ app/(service)/products/[productId]/page.tsx
@@ -1,14 +1,4 @@
-'use client';
-
-import GlobalMenu from '@/components/GlobalMenu';
-import MobileSection from '@/components/MobileSection';
-import RecentProducts from '@/components/MyPage/RecentProducts';
-import DetailInfo from '@/components/ProductDetail/DetailInfo';
-import ProductSummary from '@/components/ProductDetail/ProductSummary';
-import SizeInfo from '@/components/ProductDetail/SizeInfo';
-import Section from '@/components/Section';
-import { getProduct, Product } from '@/service/product';
-import React, { useEffect, useState } from 'react';
+import ProductDetails from "@/components/ProductDetail/ProductDetails";
 
 type Props = {
   params: {
@@ -17,27 +7,9 @@ type Props = {
 };
 
 function ProductPage({ params: { productId } }: Props) {
-  const [product, setProduct] = useState<Product>({} as Product);
-
-  useEffect(() => {
-    async function fetchProduct() {
-      const data = await getProduct(parseInt(productId));
-      setProduct(data);
-    }
-    fetchProduct();
-  }, [productId]);
-
   return (
-    <div className='mb-8 md:mb-20'>
-      <div className="hidden md:block">
-        <Section sectionName={product.productName} />
-      </div>
-      <MobileSection sectionName={product.productName} />
-      <ProductSummary product={product} />
-      <SizeInfo category={product.category} sizes={product.sizes} />
-      <DetailInfo descriptions={product.descriptions} />
-      <RecentProducts isPreview={true} />
-      <GlobalMenu />
+    <div className='mb-8'>
+      <ProductDetails productId={parseInt(productId)} />
     </div>
   );
 }
